updated some text

+++ NewDesignApp/WebServices/ApiResponse/NewApiResponse/CustomRestDetails.swift
@@ -36,9 +36,9 @@ struct CustomRestDetails: Codable {
     
     let dineUrl: String
 
-    let cateringMiles: Int
-    let largeOrder: Int
-    let largeOrderCharge: Int
+    let cateringMiles: Float
+    let largeOrder: Float
+    let largeOrderCharge: Float
     let largeOrderType: String
     let cateringNotice: String
 
@@ -46,19 +46,19 @@ struct CustomRestDetails: Codable {
     let gbDelivery: String
     let gbDeliveryCharge: Double
 
-    let deliveryMiles: Int
-    let deliveryMilesWise: Int
-    let deliveryMilesList: [Int]
+    let deliveryMiles: Float
+    let deliveryMilesWise: Float
+    let deliveryMilesList: [Float]
 
     let minDelivery: Float
-    let minPickup: Int
-    let deliveryCharge: Int
+    let minPickup: Float
+    let deliveryCharge: Float
     let deliveryChargeType: String
     let deliveryTime: Int
 
     let tax: Double
-    let conv: Int
-    let gbconv: Int
+    let conv: Float
+    let gbconv: Float
     let scharged: Float
     let schargeo: Float
     let schargev: Float
@@ -71,7 +71,7 @@ struct CustomRestDetails: Codable {
     let donateChange: String
 
     let note: String
-    let noteAmount: Int
+    let noteAmount: Float
 
     let rating: Double
     let completeMeal: [String]
