+++ NewDesignApp/WebServices/ApiResponse/NewApiResponse/RestDetailsResp.swift
@@ -43,9 +43,9 @@ struct RestDetailsRes: Codable {
        let dinein: String
        let catering: String
        let dineUrl: String
-       let cateringMiles: Int
-       let largeOrder: Int
-       let largeOrderCharge: Int
+       let cateringMiles: Float
+       let largeOrder: Float
+       let largeOrderCharge: Float
        let largeOrderType: String
        let cateringNotice: String
 
@@ -53,19 +53,19 @@ struct RestDetailsRes: Codable {
        let gbDelivery: String
        let gbDeliveryCharge: Double
 
-       let deliveryMiles: Int
-       let deliveryMilesWise: Int
-       let deliveryMilesList: [Int]
+       let deliveryMiles: Float
+       let deliveryMilesWise: Float
+       let deliveryMilesList: [Float]
 
        let minDelivery: Float
-       let minPickup: Int
-       let deliveryCharge: Int
+       let minPickup: Float
+       let deliveryCharge: Float
        let deliveryChargeType: String
        let deliveryTime: Int
 
        let tax: Double
-       let conv: Int
-       let gbconv: Int
+       let conv: Float
+       let gbconv: Float
        let scharged: Float
        let schargeo: Float
        let schargev: Float
@@ -78,7 +78,7 @@ struct RestDetailsRes: Codable {
        let donateChange: String
 
        let note: String
-       let noteAmount: Int
+       let noteAmount: Float
 
        let rating: Double
        let completeMeal: [String]
